Login refactor (#18)
* Draft login dependency

* Rename

* Add Travis Script

* Edit Ktlint

* Edit XML files (Colors, Strings, and Styles)

* Add "drawable-xhdpi" and Edit "activity_login"

* Add drawable and values files

* Update ".travis", "build.gradle" and "AndroidManifest"

* Add Fragment Package, Drawable files, and Layouts

* Update "logo_home", "activity_login", and "strings.xml"

* Create "Activity Package"

* Move "LoginActivity" to the activity package

* Update "Color Attributes"

* executable gradlew

* Update Offline Login checking

* Update LoginActivity

* Add Input Assertion

* Update Logo, Colors, Strings, Styles, and FragmentLayout

* Finish Login (Not clean yet)

* before review

* Add "View.kt" file

* Clean up "Login Feature" (No review yet)

* Merge "Going to MapActivity function" into LoginActivity

* Start MapActivity

* try fix travis

* Reformatting follow the Ktlint format

* fix android lint fatal issue

* Remove MapActivity

* Refactor LoginFragments

* Reformatting LoginFragment follow the ktlint format

* Rearrange build.gradle dependencies' alphabet

+++ ffc/src/main/java/ffc/v3/location/HouseNoLocationActivtiy.kt
@@ -35,8 +35,8 @@ import ffc.v3.R
 import ffc.v3.api.FfcCentral
 import ffc.v3.api.PlaceService
 import ffc.v3.util.org
-import kotlinx.android.synthetic.main.activity_house_no_location.*
-import kotlinx.android.synthetic.main.item_house.view.*
+import kotlinx.android.synthetic.main.activity_house_no_location.houseList
+import kotlinx.android.synthetic.main.item_house.view.houseNo
 import org.jetbrains.anko.bundleOf
 import org.jetbrains.anko.defaultSharedPreferences
 import org.jetbrains.anko.find
@@ -68,7 +68,7 @@ class HouseNoLocationActivtiy : BaseActivity() {
         })
 
         val org = defaultSharedPreferences.org!!
-        FfcCentral().service<PlaceService>().listHouseNoLocation(org.id.toLong()).enqueue {
+        FfcCentral().service<PlaceService>().listHouseNoLocation(org.id).enqueue {
             onSuccess {
                 if (BuildConfig.DEBUG) {
                     toast("Loaded ${body()?.size}")
@@ -150,7 +150,8 @@ class HouseNoLocationActivtiy : BaseActivity() {
             }
 
         override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HouseViewHolder {
-            val view = LayoutInflater.from(parent.context).inflate(R.layout.item_house, parent, false)
+            val view = LayoutInflater.from(parent.context)
+                .inflate(R.layout.item_house, parent, false)
             return HouseViewHolder(view, onItemClick)
         }
