+++ ffc/src/main/java/ffc/v3/location/MarkLocationActivity.kt
@@ -35,7 +35,7 @@ import ffc.v3.api.FfcCentral
 import ffc.v3.api.PlaceService
 import ffc.v3.util.find
 import ffc.v3.util.toPoint
-import kotlinx.android.synthetic.main.activity_add_location.*
+import kotlinx.android.synthetic.main.activity_add_location.done
 import org.jetbrains.anko.alert
 import org.jetbrains.anko.indeterminateProgressDialog
 import org.jetbrains.anko.startActivityForResult
@@ -88,7 +88,7 @@ class MarkLocationActivity : BaseActivity() {
     private fun updateHouse() {
         val dialog = indeterminateProgressDialog("updating")
 
-        FfcCentral().service<PlaceService>().updateHouse(org.id.toLong(), targetPlace).enqueue {
+        FfcCentral().service<PlaceService>().updateHouse(org.id, targetPlace).enqueue {
             always { dialog.dismiss() }
             onSuccess {
                 setResult(Activity.RESULT_OK)
